Butter Knife, Dagger, LoginActivity, a bit of MVP
This commit does quite some refactoring. It makes some use of Butter Knife
and Dagger to simplify the usage of singletons and to remove the all the
findViewById calls.

Then, a simple LoginActivity has been implemented (but not been tested
yet) which will establish the connection and go through the actual login
process.

I also started implemented the MVP pattern a bit in the MainActivity but
totally forgot about it again in the LoginActivity. I still have to learn
a lot.