Actually implement Dagger
Previously there was no injection so the app would've just crashed.
However, I didn't try it out so yep, didn't really find that out until
later when I looked at a guide again.

The API methods still miss a lot of error handling. For example, the user
will be told they're logged in because the API returns an actual SOAP
response with the error code for the wrong password and doesn't just set
the HTTP status code to 403 or something like that.