Add SessionKeepAliveResponse class
This class is used to deserialize the SessionKeepAliveResponse which
should follow a SessionKeepAlive request. It works the same way the
LoginResponse works, too.

The SessionKeepAliveResponse tag inside of the SOAP body usually is empty
here, too, meaning we want to convert it to an empty string using our
custom converter.

The rest of the class works exactly the way the LoginResponse works, too,
thanks to the EasyBox following an actual standard.