Let Dagger provide the ApplicationSettings
The ApplicationModule was changed so it would also provide an
implementation of the ApplicationSettings whenever needed. For this, the
application context is needed.

Of course, the settings could've been moved to their own module but in my
opinion this didn't really make sense - they are the *application*
settings after all.

In addition, a new ApplicationComponent has been added which can be
injected into the two activities.