Fix login error handling in LoginPresenter
Previously the app would crash when using a wrong password when trying to
log in. In that case, the faultMsgCode would be null and thus the regex
matching would throw an exception (can't match anything on null). This has
been fixed by making sure the message code isn't null before trying to
perform a regex match on it.